Use basic-range datepicker for pause reminders
