+++ frontend/src/app/shared/components/datepicker/basic-range-date-picker/basic-range-date-picker.component.html
@@ -1,7 +1,8 @@
 <input
   #input
   type="text"
-  class="spot-input"
+  class="spot-input op-basic-range-datepicker--input"
+  data-qa-selector="op-basic-range-date-picker"
   [ngClass]="inputClassNames"
   [attr.data-value]="value"
   [id]="id"
