+++ frontend/src/app/shared/components/datepicker/basic-range-date-picker/basic-range-date-picker.component.sass
@@ -1,0 +1,3 @@
+.op-basic-range-datepicker
+  &--input
+    min-width: 12rem
