+++ spec/support/components/datepicker/range_datepicker.rb
@@ -1,0 +1,7 @@
+require_relative './basic_datepicker'
+
+module Components
+  class RangeDatepicker < BasicDatepicker
+    include MonthRangeSelection
+  end
+end
